#180 - `TodayPopularPostingService` 서비스 클래스의 순환참조 문제 해결
- `TodayPopularPostingService` 서비스 클래스에서 인기 점수와 인기 게시글을 얻어오는 기능을 구현하면서 PostingService, CommentService와 같은 클래스와 순환참조 문제가 발생했다.
- 이를 해결하기 위해 인기 점수를 관리하는 서비스인 `TodayPopularScoreService`을 따로 만들어주었다.
- 인기 리뷰도 이와 같이 분리할 예정이다.
