+++ src/main/java/com/ncookie/imad/domain/posting/service/PostingService.java
@@ -12,7 +12,7 @@ import com.ncookie.imad.domain.posting.repository.PostingRepository;
 import com.ncookie.imad.domain.profile.entity.PostingScrap;
 import com.ncookie.imad.domain.profile.service.ScrapService;
 import com.ncookie.imad.domain.ranking.service.ContentsRankingScoreUpdateService;
-import com.ncookie.imad.domain.ranking.service.TodayPopularPostingService;
+import com.ncookie.imad.domain.ranking.service.TodayPopularScoreService;
 import com.ncookie.imad.domain.user.entity.UserAccount;
 import com.ncookie.imad.global.Utils;
 import com.ncookie.imad.domain.user.service.UserRetrievalService;
@@ -47,7 +47,7 @@ public class PostingService {
     private final ScrapService scrapService;
 
     private final ContentsRankingScoreUpdateService contentsRankingScoreUpdateService;
-    private final TodayPopularPostingService todayPopularPostingService;
+    private final TodayPopularScoreService todayPopularScoreService;
 
 
     public PostingDetailsResponse getPosting(String accessToken, Long postingId) {
@@ -84,7 +84,7 @@ public class PostingService {
         // TODO: 특정 기준(accessToken에 조회 여부 저장 등)을 통해 중복 조회수를 필터링 해야함
         // 조회수 갱신
         postingRepository.updateViewCount(postingId, posting.getViewCnt() + 1);
-        todayPopularPostingService.addPopularScore(posting, TodayPopularPostingService.POPULAR_VIEW_CNT_SCORE);
+        todayPopularScoreService.addPopularScore(posting, TodayPopularScoreService.POPULAR_VIEW_CNT_SCORE);
 
         // 게시글 정보, 댓글 리스트, like status 등을 DTO 객체에 저장
         PostingDetailsResponse postingDetailsResponse = PostingDetailsResponse.toDTO(posting, commentList);
@@ -308,7 +308,7 @@ public class PostingService {
             // 기존의 좋아요를 좋아요 취소 또는 싫어요로 변경했을 때 인기 점수 차감
             int previousLikeStatus = postingLike.getLikeStatus();
             if (previousLikeStatus == 1 && (likeStatus == 0 || likeStatus == -1)) {
-                todayPopularPostingService.subtractPopularScore(posting, TodayPopularPostingService.POPULAR_LIKE_SCORE);
+                todayPopularScoreService.subtractPopularScore(posting, TodayPopularScoreService.POPULAR_LIKE_SCORE);
             }
 
             // like_status가 1이면 좋아요, -1이면 싫어요, 0이면 둘 중 하나를 취소한 상태이므로 테이블에서 데이터 삭제
@@ -327,7 +327,7 @@ public class PostingService {
 
         // 게시글 인기 점수 추가
         if (likeStatus == 1) {
-            todayPopularPostingService.addPopularScore(posting, TodayPopularPostingService.POPULAR_LIKE_SCORE);
+            todayPopularScoreService.addPopularScore(posting, TodayPopularScoreService.POPULAR_LIKE_SCORE);
         }
 
         // like, dislike count 갱신
