+++ src/main/java/com/ncookie/imad/domain/profile/service/ProfileService.java
@@ -10,7 +10,7 @@ import com.ncookie.imad.domain.profile.dto.response.*;
 import com.ncookie.imad.domain.profile.entity.ContentsBookmark;
 import com.ncookie.imad.domain.profile.entity.PostingScrap;
 import com.ncookie.imad.domain.ranking.service.ContentsRankingScoreUpdateService;
-import com.ncookie.imad.domain.ranking.service.TodayPopularPostingService;
+import com.ncookie.imad.domain.ranking.service.TodayPopularScoreService;
 import com.ncookie.imad.domain.review.dto.response.ReviewListResponse;
 import com.ncookie.imad.domain.review.service.ReviewService;
 import com.ncookie.imad.domain.user.entity.UserAccount;
@@ -48,7 +48,7 @@ public class ProfileService {
     private final ScrapService scrapService;
 
     private final ContentsRankingScoreUpdateService contentsRankingScoreUpdateService;
-    private final TodayPopularPostingService todayPopularPostingService;
+    private final TodayPopularScoreService todayPopularScoreService;
 
 
     // 프로필 요약 정보
@@ -165,7 +165,7 @@ public class ProfileService {
             log.info("스크랩 등록 완료");
 
             contentsRankingScoreUpdateService.addRankingScore(posting.getContents(), SCRAP_RANKING_SCORE);
-            todayPopularPostingService.addPopularScore(posting, TodayPopularPostingService.POPULAR_SCRAP_SCORE);
+            todayPopularScoreService.addPopularScore(posting, TodayPopularScoreService.POPULAR_SCRAP_SCORE);
             log.info("[게시글 스크랩 추가] 랭킹 점수 반영 완료");
 
             return ResponseCode.SCRAP_ADD_SUCCESS;
@@ -186,7 +186,7 @@ public class ProfileService {
         }
         PostingScrap postingScrap = scrapService.findByIdAndUserAccount(scrapId, user);
         contentsRankingScoreUpdateService.subtractRankingScore(postingScrap.getPosting().getContents(), SCRAP_RANKING_SCORE);
-        todayPopularPostingService.subtractPopularScore(postingScrap.getPosting(), TodayPopularPostingService.POPULAR_SCRAP_SCORE);
+        todayPopularScoreService.subtractPopularScore(postingScrap.getPosting(), TodayPopularScoreService.POPULAR_SCRAP_SCORE);
         log.info("[게시글 스크랩 취소] 랭킹 점수 반영 완료");
 
         scrapService.deleteByIdAndUserAccount(scrapId, user);
