+++ src/main/java/com/ncookie/imad/domain/ranking/service/TodayPopularPostingService.java
@@ -1,7 +1,6 @@
 package com.ncookie.imad.domain.ranking.service;
 
 import com.ncookie.imad.domain.posting.dto.response.PostingDetailsResponse;
-import com.ncookie.imad.domain.posting.entity.Posting;
 import com.ncookie.imad.domain.posting.service.PostingService;
 import com.ncookie.imad.domain.ranking.entity.TodayPopularPosting;
 import com.ncookie.imad.domain.ranking.repository.TodayPopularPostingsRepository;
@@ -12,7 +11,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 
 @Slf4j
@@ -23,12 +21,6 @@ public class TodayPopularPostingService {
     private final PostingService postingService;
 
     private final TodayPopularPostingsRepository todayPopularPostingsRepository;
-
-    public final static int POPULAR_SCRAP_SCORE = 10;
-    public final static int POPULAR_LIKE_SCORE = 5;
-    public final static int POPULAR_COMMENT_SCORE = 2;
-    public final static int POPULAR_VIEW_CNT_SCORE = 1;
-
 
     public PostingDetailsResponse getTodayPopularPosting(String accessToken) {
         List<TodayPopularPosting> popularPostingList = todayPopularPostingsRepository.findTopByPopularScore();
@@ -49,59 +41,5 @@ public class TodayPopularPostingService {
         
         log.info("오늘의 게시글 데이터를 반환합니다");
         return postingService.getPosting(accessToken, popularPostingList.get(0).getPosting().getPostingId());
-    }
-
-    public void addPopularScore(Posting posting, int score) {
-        if (posting == null) {
-            log.warn("인기 점수 반영 실패 : ID에 해당하는 리뷰가 존재하지 않음");
-            return;
-        }
-
-        Optional<TodayPopularPosting> popularPostingOptional = todayPopularPostingsRepository.findByPosting(posting);
-        if (popularPostingOptional.isPresent()) {
-            TodayPopularPosting todayPopularPosting = popularPostingOptional.get();
-            Long oldScore = todayPopularPosting.getPopularScore();
-
-            todayPopularPosting.setPopularScore(oldScore + score);
-            todayPopularPostingsRepository.save(todayPopularPosting);
-        } else {
-            todayPopularPostingsRepository.save(
-                    TodayPopularPosting.builder()
-                            .posting(posting)
-                            .popularScore((long) score)
-                            .build()
-            );
-        }
-        
-        log.info("게시글 인기 점수 추가 완료");
-    }
-
-    public void subtractPopularScore(Posting posting, int score) {
-        if (posting == null) {
-            log.warn("인기 점수 반영 실패 : ID에 해당하는 리뷰가 존재하지 않음");
-            return;
-        }
-
-        Optional<TodayPopularPosting> popularPostingOptional = todayPopularPostingsRepository.findByPosting(posting);
-        if (popularPostingOptional.isPresent()) {
-            TodayPopularPosting todayPopularPosting = popularPostingOptional.get();
-            Long oldScore = todayPopularPosting.getPopularScore();
-
-            todayPopularPosting.setPopularScore(oldScore - score);
-            todayPopularPostingsRepository.save(todayPopularPosting);
-        } else {
-            todayPopularPostingsRepository.save(
-                    TodayPopularPosting.builder()
-                            .posting(posting)
-                            .popularScore((long) score)
-                            .build()
-            );
-        }
-
-        log.info("게시글 인기 점수 추가 완료");
-    }
-
-    public void clearDaily() {
-        todayPopularPostingsRepository.deleteAllInBatch();
     }
 }
