+++ src/main/java/com/ncookie/imad/domain/ranking/service/TodayPopularScheduler.java
@@ -11,21 +11,20 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 @Component
 public class TodayPopularScheduler {
-    private final TodayPopularReviewService todayPopularReviewService;
-    private final TodayPopularPostingService todayPopularPostingService;
+    private final TodayPopularScoreService todayPopularScoreService;
 
 
     @Description("매일 자정마다 인기 리뷰 데이터 리셋")
     @Scheduled(cron = "0 0 0 * * ?")    // 자정마다 실행
     public void clearPopularReviewDaily() {
-        todayPopularReviewService.clearDaily();
+        todayPopularScoreService.clearDaily();
         log.info("오늘의 리뷰 데이터 리셋 완료");
     }
 
     @Description("매일 자정마다 인기 게시글 데이터 리셋")
     @Scheduled(cron = "0 0 0 * * ?")    // 자정마다 실행
     public void clearPopularPostingDaily() {
-        todayPopularPostingService.clearDaily();
+        todayPopularScoreService.clearDaily();
         log.info("오늘의 게시글 데이터 리셋 완료");
     }
 }
